test: consolidate story tests
